save lab result on backend

+++ htdocs/frontend/controllers/LabController.php
@@ -5,6 +5,7 @@ namespace frontend\controllers;
 use common\models\Component;
 use common\models\Lab;
 use common\models\LabItems;
+use common\models\LabResults;
 use common\models\Student;
 use common\models\User;
 use Jurosh\PDFMerge\PDFMerger;
@@ -14,6 +15,8 @@ use yii\helpers\Json;
 use yii\web\Controller;
 use yii\filters\AccessControl;
 use yii\web\Response;
+use yii\web\NotFoundHttpException;
+
 
 class LabController extends Controller
 {
@@ -65,12 +68,38 @@ class LabController extends Controller
     /**
      * @param $number
      * @return string
+     * @throws NotFoundHttpException
      */
     public function actionLab($number)
     {
         $session = Yii::$app->session;
         if ($session->has('lab_number')) {
             $session->remove('lab_number');
+        }
+
+        $student = Student::find()->andWhere(['user_id' => Yii::$app->user->id])->one();
+
+        if (!$student->group->labs->{"lab$number"}) {
+            throw new NotFoundHttpException('Лабораторная работа недоступна');
+        }
+
+        $studentLabs = $student->labs;
+
+        /** @var LabResults $labResult */
+        if ($labResult = $studentLabs->{"lab$number"}) {
+            if ($labResult->success) {
+                throw new NotFoundHttpException('Лабораторная работа уже выполнена');
+            }
+
+            $labResult->attempts = $labResult->attempts + 1;
+            $labResult->save();
+        } else {
+            $labResult = new LabResults();
+            $labResult->attempts = 1;
+            $labResult->success = false;
+            $labResult->save();
+            $studentLabs->{"lab$number" . "_id"} = $labResult->id;
+            $studentLabs->save();
         }
 
         $session->set('lab_number', $number);
@@ -159,41 +188,69 @@ class LabController extends Controller
         $session = Yii::$app->session;
         $request = Json::decode(Yii::$app->getRequest()->getRawBody());
 
+
         if ($session->has('lab_number')) {
-            $lab = Lab::findOne($session->get('lab_number'));
+            $labNumber = $session->get('lab_number');
+            $student = Student::find()->andWhere(['user_id' => Yii::$app->user->id])->one();
 
-            if (isset($request['taskPdf']) && isset($request['titlePdf'])) {
-                $user = User::find()->andWhere(['id' => Yii::$app->user->id])->one();
-                $fileName = $user->username . '_' . $lab->id . '_' . time();
+            if ($student->group->labs->{"lab$labNumber"}) {
+                /** @var LabResults $labResult */
+                if ($labResult = $student->labs->{"lab$labNumber"}) {
 
-                $taskPdfBase64 = htmlspecialchars($request['taskPdf']);
-                $titlePdfBase64 = htmlspecialchars($request['titlePdf']);
+                    if (!$labResult->success) {
+                        $lab = Lab::findOne($labNumber);
+                        $session->remove('lab_number');
 
-                $titlePdf = base64_decode($titlePdfBase64);
-                $taskPdf = base64_decode($taskPdfBase64);
+                        if (isset($request['taskPdf']) && isset($request['titlePdf'])) {
+                            $fileName = $student->user->username . '_' . $lab->id . '_' . time();
 
-                $titlePath = __DIR__ . "/tmp/title_$fileName.pdf";
-                $taskPath = __DIR__ . "/tmp/task_$fileName.pdf";
+                            $taskPdfBase64 = htmlspecialchars($request['taskPdf']);
+                            $titlePdfBase64 = htmlspecialchars($request['titlePdf']);
 
-                $resultPath = $_SERVER['DOCUMENT_ROOT'] . "../../../data/results/$fileName.pdf";
+                            $titlePdf = base64_decode($titlePdfBase64);
+                            $taskPdf = base64_decode($taskPdfBase64);
 
-                file_put_contents($titlePath, $titlePdf);
-                file_put_contents($taskPath, $taskPdf);
+                            $titlePath = __DIR__ . "/tmp/title_$fileName.pdf";
+                            $taskPath = __DIR__ . "/tmp/task_$fileName.pdf";
 
-                $pdf = new PDFMerger();
+                            $resultPath = $_SERVER['DOCUMENT_ROOT'] . "../../../data/results/$fileName.pdf";
 
-                $pdf->addPDF($titlePath, 'all', 'vertical')->addPDF($taskPath, 'all', 'vertical');
-                $pdf->merge('file', $resultPath);
+                            file_put_contents($titlePath, $titlePdf);
+                            file_put_contents($taskPath, $taskPdf);
 
-                unlink($titlePath);
-                unlink($taskPath);
+                            $pdf = new PDFMerger();
 
-                Yii::$app->response->statusCode = 200;
-                return [
-                    'date' => date('d.m.Y'),
-                    'file_path' => "/data/results/$fileName.pdf",
-                    'attempt'=> 2
-                ];
+                            $pdf->addPDF($titlePath, 'all', 'vertical')->addPDF($taskPath, 'all', 'vertical');
+                            $pdf->merge('file', $resultPath);
+
+                            unlink($titlePath);
+                            unlink($taskPath);
+
+                            $labResult->success = true;
+                            $labResult->created_at = time();
+                            $labResult->file_path = "/data/results/$fileName.pdf";
+
+                            if ($labResult->save()) {
+                                Yii::$app->response->statusCode = 200;
+                                return [
+                                    'date' => date('d.m.Y i:H', $labResult->created_at),
+                                    'file_path' => $labResult->file_path,
+                                    'attempt' => $labResult->attempts
+                                ];
+                            } else {
+                                return ['error' => 'Произошла ошибка при сохранении отчета'];
+                            }
+                        } else {
+                            return ['error' => 'Произошла ошибка при сохранении отчета'];
+                        }
+                    } else {
+                        return ['error' => 'Работа уже выполнена'];
+                    }
+                } else {
+                    return ['error' => 'Работа не найдена'];
+                }
+            } else {
+                return ['error' => 'Работа недоступна'];
             }
         }
